refactor: 피드 상세 조회 로직 구현클래스로 이동
